% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dna_convert}
\alias{dna_convert}
\title{Convert image DNA to PNG format}
\usage{
dna_convert(dna, maxXY, tempf, pngWH, bg = "white")
}
\arguments{
\item{dna}{matrix or character, untangled or tangled image DNA of any size.}

\item{tempf}{temporate file generated by default or given as file path.}

\item{pngWH}{vector, width and height of reconstructed image. If missing, width and height of original image are used.}

\item{bg}{character, color or RGB code indicating the background color of PNG.}
}
\description{
Function converts image DNA to array object including RGB or gray scale for each pixel.
}
\details{
See example...
}
\examples{
dna <- dna_untangle(dna_in(rgb = FALSE))
for(i in 1:20){
dna <- dna_mutate(dna)
}
test <- dna_convert(dna)
grid::grid.raster(test)
test[1,1,]
}
