% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dna_in}
\alias{dna_in}
\title{Generate image DNA}
\usage{
dna_in(n = 10, vertex = 3, maxXY = c(200, 200), rgb = TRUE)
}
\arguments{
\item{n}{numeric, number of polygons to be drawn to your image, Default = 10.}

\item{vertex}{numeric, number of vertices each polygen should be constructed of, Default = 3.}

\item{maxXY}{vector, sets the pixels given by the original image, Default = c(200, 200).}

\item{rgb}{logic, if TRUE, generates an RGB code (default), if FALLS, generates a gray scale.}
}
\description{
Function to generate a random image DNA.
}
\details{
Is generally the first step of starting an image reproduction. Use maxXY = dim(pic)[2:1] to get the number of pixels in the original image.
}
\examples{
dna_in(n = 5, vertex = 2, rgb = FALSE)
}
