% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dna_print}
\alias{dna_print}
\title{Plot image DNA}
\usage{
dna_print(dna, maxXY)
}
\arguments{
\item{dna}{matrix or character, untangled or tangled image DNA of any size.}
}
\description{
Function to plot an image DNA.
}
\details{
Are comming up soon...
}
\examples{
#Is comming up soon...
}
