% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dna_read}
\alias{dna_read}
\title{Read image DNA}
\usage{
dna_read(log, tangle = TRUE)
}
\arguments{
\item{log}{filepath.}

\item{tangle}{logic, TRUE if tangled (default), FALSE if untangled.}
}
\description{
Function to read file.txt including a tangled or untangled image DNA.
}
\details{
After your calculation is done, you may want to improve the results by reloading the existing image DNA, and start the calculation again.
}
