% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dna_tangle}
\alias{dna_tangle}
\title{Tangle image DNA}
\usage{
dna_tangle(dna)
}
\arguments{
\item{dna}{matrix, untangled image DNA of any size.}
}
\description{
Function to tangle an untangled image DNA.
}
\details{
See example...
}
\examples{
dna <- dna_untangle(dna_in(n = 2, vertex = 3, rgb = FALSE))
dna_tangle(dna)
}
