% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dna_untangle}
\alias{dna_untangle}
\title{Untangle image DNA}
\usage{
dna_untangle(dna)
}
\arguments{
\item{dna}{character, tangled image DNA of any size.}
}
\description{
Function to untangle an image DNA to get color code and coordinates of each polygon.
}
\details{
See example...
}
\examples{
dna_untangle(dna_in())
dna_untangle(dna_in(rgb = FALSE))
}
