% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gen_art}
\alias{gen_art}
\title{Genetic algorithm}
\usage{
gen_art(pic, dna, iter = Inf, fname = "ReconPic.png",
  fnamelog = "Log.txt", tempf, degree = "Soft")
}
\arguments{
\item{pic}{object of class array, original image in PNG format.}

\item{dna}{matrix or character, untangled or tangled image DNA of any size.}

\item{iter}{numeric, number of iterations. Default = Inf.}

\item{fname}{character, filename for the reconstructed image. Default = "ReconPic.png".}

\item{tempf}{temporate file generated by default or given as file path. See ...}

\item{degree}{numeric, how many nucleotides should be modified per one iteration, Default = 1.}

\item{fanmelog}{character, filename for file.txt including the tangled image DNA. Default = "Log.txt".}
}
\description{
Algorithm that continues to mutate (improve) the image DNA until manually stoped or a given maximum of iterations is reached
}
\details{
It is recommended to use the softest mutation rate as given by default.

See example...
}
\examples{
#Is coming up soon...
}
