% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pic_compare}
\alias{pic_compare}
\title{Calculate image fitness}
\usage{
pic_compare(imgONE, imgTWO)
}
\arguments{
\item{imgONE}{first array to compare with second array.}

\item{imgTWO}{second array to compare with first array.}
}
\description{
Function to compare two images to calculate fitness.
}
\details{
...
}
