% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pic_convert}
\alias{pic_convert}
\title{Convert RGB code to gray scale}
\usage{
pic_convert(pic)
}
\arguments{
\item{pic}{array including RGB.}
}
\description{
Function to convert RGB to gray scale.
}
\details{
...
}
