% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pic_randomtiles}
\alias{pic_randomtiles}
\title{Randomly crop an image}
\usage{
pic_randomtiles(w, h, plot = TRUE)
}
\arguments{
\item{w}{numeric, width of the original picture.}

\item{h}{numeric, height of the original picture.}

\item{plot}{logic, function plots layout if TRUE (default).}
}
\description{
Function to tile an image into 4 different pieces.
}
\details{
Try multiple times to get the required image layout. Assign data frame to an object for subsequent use in gen_art. See example.
}
\examples{
## Image width and height
h <- 100
w <- 120

## Three trials as an example
for(i in 1:3){
   pic_randomtiles(w, h, plot = TRUE)
   title("Split your image for better results...")
   text(w/2, h/2, paste("Trial", i), cex = 4, col = "white")
   Sys.sleep(1.5)
}
}
