% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pic_unitiles}
\alias{pic_unitiles}
\title{Uniformly crop an image}
\usage{
pic_unitiles(wh, nn, plot = TRUE)
}
\arguments{
\item{wh}{vector, width and height of the original image.}

\item{nn}{vector, number of columns and rows to split the image in.}

\item{plot}{logic, function plots layout if TRUE (default).}
}
\description{
Function to uniformly crop an image into n pieces.
}
\details{
Function cuts remaining pixels if width/height is indivisible by columns/rows. Assign data frame to an object for subsequent use in gen_art. See example.
}
\examples{
## Image width/height and number of colums/rows
wh <- c(89, 70)
nn <- c(6, 4)

## Calculation and plotting
pic_unitiles(wh, nn)
}
